mcp-memory-service/requirements.txt
# FastAPI para API REST
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.8.0

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pyyaml==6.0.1

# Logging
structlog==23.2.0 